refactor: proper name for apiRequests.ts
